+++ src/app/modules/observation-code-lookup/observation-code-lookup.component.ts
@@ -309,6 +309,7 @@ export class ObservationCodeLookupComponent
                         if (contains.length > count) {
                           contains.length = count;
                         }
+                        this.appendCodeSystemToDuplicateDisplay(contains);
                         // Emit a complete notification
                         return EMPTY;
                       }
@@ -425,29 +426,36 @@ export class ObservationCodeLookupComponent
     return (bundle.entry || []).reduce((acc, entry) => {
       const observation = entry.resource as Observation;
       const datatype = this.getValueDataType(observation);
-      if (
-        this.currentData.datatype === ANY_DATATYPE ||
-        !this.currentData.datatype ||
-        datatype === this.currentData.datatype
-      ) {
-        acc.push(
-          ...observation.code.coding
-            .filter((coding) => {
-              const matched =
-                !processedCodes[coding.code] &&
-                selectedCodes.indexOf(coding.code) === -1;
+      acc.push(
+        ...(observation.code.coding
+          ?.filter((coding) => {
+            let matched = false;
+            if (coding.code && !processedCodes[coding.code]) {
+              // Even though this observation's data type does not match selected codes, we want to
+              // go through its codings and mark 'processedCodes' accordingly, so that these codes
+              // can be excluded from next queries. Otherwise, we might get into a near-infinite loop
+              // of queries returning the same code. This happened with searching "Total Cholesterol"
+              // and selecting code "14647-2".
               processedCodes[coding.code] = true;
-              return matched;
-            })
-            .map((coding) => {
-              this.code2Type[coding.system + '|' + coding.code] = datatype;
-              return {
-                code: { code: coding.code, system: coding.system },
-                display: coding.display
-              };
-            })
-        );
-      }
+              if (
+                (!this.currentData.datatype ||
+                  this.currentData.datatype === ANY_DATATYPE ||
+                  datatype === this.currentData.datatype) &&
+                selectedCodes.indexOf(coding.code) === -1
+              ) {
+                matched = true;
+              }
+            }
+            return matched;
+          })
+          .map((coding) => {
+            this.code2Type[coding.system + '|' + coding.code] = datatype;
+            return {
+              code: { code: coding.code, system: coding.system },
+              display: coding.display || coding.code
+            };
+          }) || [])
+      );
       return acc;
     }, []);
   }
@@ -517,5 +525,26 @@ export class ObservationCodeLookupComponent
     } else {
       this.elementRef.nativeElement.removeAttribute('aria-describedby');
     }
+  }
+
+  /**
+   * For autocomplete items with the same display and different code + code system
+   * combination, append code + code system to the display so distinct items are
+   * shown to user.
+   * @param contains the array of items for the autocomplete
+   */
+  appendCodeSystemToDuplicateDisplay(contains: any[]): void {
+    // an array of displays that have more than one appearance.
+    const duplicateDisplays = contains
+      .filter(
+        (item, index, arr) =>
+          arr.findIndex((x) => x.display === item.display) !== index
+      )
+      .map((item) => item.display);
+    contains.forEach((item) => {
+      if (duplicateDisplays.includes(item.display)) {
+        item.display = `${item.display} | ${item.code.code} | ${item.code.system}`;
+      }
+    });
   }
 }